fix parsing debug location with Windows-style absolute path
Debug locations with Windows-style absolute path were incorrectly parsed
because of the ':' character following the drive letter, which is also
used as the delimiter before line and column numbers. The issue is fixed
by scanning the string in reverse order, so that anything before the
line number and the column number is treated as a path.